Move message methods to its own file
